+++ changelog.md
@@ -117,5 +117,6 @@
 
 ## Next
 
+* New users must initialize account before access granted
 * textSpaced form field format - which preserves line spacing
 * Code tidy